requirement.txt
langchain
streamlit
python-dotenv
PyPDF2 
google-generativeai 
langchain-google-genai
langchain-community


-e .